requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy
openvino>=2023.2.0
onnx>=1.11.0
git+https://github.com/huggingface/transformers.git
git+https://github.com/huggingface/optimum-intel.git
semantic-kernel==0.4.3.dev0
sentence-transformers
scikit-learn
jupyterlab
ipywidgets
ipykernel>=5.*
ipython>=7.16.3